+++ code-gen/templates/p1as-observability/config.sh
@@ -1,3 +1,3 @@
 export CDE_DEPLOY=true
-export CHUB_DEPLOY=true
+export CHUB_DEPLOY=("p1as-cloudwatch" "p1as-newrelic" "p1as-grafana" "p1as-prometheus" "p1as-prometheus-agent")
 export PRIMARY_REGION_ONLY_DEPLOY=("p1as-grafana" "p1as-prometheus")